+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Debug console supports `__modname__` prefix while in break
 * Better error reporting for tasks.json tasks
+* Branch name configuration for Publish was not correctly applied
 
 ## 0.18.40